updates to follow App Template rules
- update name, summary and description
- update files to use shorter name "CoinGecko Price"
